ObjectPoolManagerTest.cs GameObject 기반 풀링으로 수정완료
현재 Monster에 IPoolObject 부착하여 Addressable에 Init_Pool 라벨이 정상적으로 풀링 되는 것은 확인했으나 중간에 몬스터 SO 지정이 다른게 있어서 Content 쪽에서 수정 후 확인 예정
일단 풀링 성공하는 것은 확인 완료

이후 StageManager와 MonsterSpawner를 PoolManager와 연동할 듯?

+++ Assets/03.Data/05.Save/BaseRequestData.cs
@@ -98,8 +98,8 @@ public class UserProfileRequest : BaseRequestData
     public float currentExp;
     public int currentStage;
     public long lastLoginTimestamp;
-    public long gold;
-    public long dia;
+    public double gold;
+    public double dia;
 
     // UpgradeSystem 연동 데이터
     public Dictionary<string, int> upgradeTrackLevels = new();
